requirements.txt
git+https://github.com/huggingface/transformers
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
matplotlib
numpy
Pillow
accelerate
tokenizers
datasets
requests
scipy
scikit-image